wip Event system
Plugins will listen for and emit events. The "state" of the game is the
sum of the separate plugins' states.

The specific types here are a quick and dirty outline, but this will
be refined a lot as we build a few plugins and get a sense of what
kind of cross-plugin interactions and state are necessary. The overall
design is inspired by semantic web ideas and DDD.